Made changes to allow users to edit or delete only questions and answers they have created

+++ resources/views/answer.blade.php
@@ -12,12 +12,17 @@
                     </div>
 
                     <div class="card-footer">
-                        {{ Form::open(['method'  => 'DELETE', 'route' => ['answers.destroy', $question, $answer->id]])}}
-                        <button class="btn btn-danger float-right mr-2" value="submit" type="submit" id="submit">Delete</button>
-                        {!! Form::close() !!}
-                        <a class="btn btn-primary float-right" href="{{ route('answers.edit',['question_id'=> $question, 'answer_id'=> $answer->id, ])}}">
-                            Edit Answer
-                        </a>
+                        @if(Auth::user() == $answer->user)
+                            {{ Form::open(['method'  => 'DELETE', 'route' => ['answers.destroy', $question, $answer->id]])}}
+                            <button class="btn btn-danger float-right mr-2" value="submit" type="submit" id="submit">
+                                Delete
+                            </button>
+                            {!! Form::close() !!}
+                            <a class="btn btn-primary float-right"
+                               href="{{ route('answers.edit',['question_id'=> $question, 'answer_id'=> $answer->id, ])}}">
+                                Edit Answer
+                            </a>
+                        @endif
                     </div>
                 </div>
